bug: fixed
* Changing user image was not updating

* BREAKING change user image on browser also

+++ test.py
@@ -14,13 +14,22 @@ def main(page: ft.Page) -> None:
     ft_image = ft.Image(src='uploads')
 
     def upload_files(e: ft.FilePickerUploadFile):
-
-        for x in MyPickFiles.result.files:
-            if os.path.exists(x.name):
-                os.remove(x.name)
-            dest = os.path.join(os.getcwd(), "uploads")
-            shutil.copy(x.path, f'{dest}')
-            ft_image.src = f'uploads/{x.name}'
+        if page.web:
+            uf = []
+            f: ft.FilePickerUploadFile
+            for f in MyPickFiles.result.files:
+                uf.append(ft.FilePickerUploadFile(f.name, upload_url=page.get_upload_url(f.name, 600)))
+            MyPickFiles.upload(uf)
+            for f in MyPickFiles.result.files:
+                ft_image.src = f'/uploads/{f.name}'
+            page.update()
+        else:
+            for x in MyPickFiles.result.files:
+                if os.path.exists(x.name):
+                    os.remove(x.name)
+                dest = os.path.join(os.getcwd(), "assets/uploads")
+                shutil.copy(x.path, f'{dest}')
+                ft_image.src = f'/uploads/{x.name}'
 
             page.update()
 
@@ -42,4 +51,4 @@ def main(page: ft.Page) -> None:
     )
 
 
-ft.app(main)
+ft.app(main, assets_dir='assets', upload_dir='assets/uploads', view=ft.AppView.WEB_BROWSER)
